+++ CHANGES.md
@@ -9,6 +9,7 @@
 - Add `min_width` parameter to `AlignTestCases` transformer. It allows to set minimal width of data column ([#242](https://github.com/MarketSquare/robotframework-tidy/issues/242))
 - Changed behaviour of `DiscardEmptySections` transformer. Now by default section with only comments are not considered empty. 
   Configure `allow_only_comments` to `False` if you want to remove those sections ([#235](https://github.com/MarketSquare/robotframework-tidy/issues/235))
+- Add `skip_types` parameter to `AlignVariablesSection` which allows to not align variables of particular type ([#225](https://github.com/MarketSquare/robotframework-tidy/issues/225))
 
 ## 1.6.2